Hotfix for the "object is not iterable" issue.
Hotfix for the "object is not iterable" issue. The ideal todo here would be to refactor and standardize what we're expecting. Is it an object (non iterable) or an array? Maybe looping over the objects in an older / more standard way would be better here (a.k.a. Object.keys(), Object.entries(), etc.)